Add files via upload
Middle finger and toothbrush sized to fit inside of the cheap plastic snap-together eggs. STL files are in the zip folder.  I printed the hand at .16mm layer height and the toothbrush with .20mm layer height. 5% infill for both. No supports needed. I used an 8mm brim on the hand.